+++ app/renderer/js/pages/preference/general-section.ts
@@ -91,7 +91,7 @@ export function initGeneralSection({$root}: GeneralSectionProperties): void {
         </div>
         <div class="setting-row" id="silent-option">
           <div class="setting-description">
-            ${t.__("Mute all sounds from Zulip")}
+            ${t.__("Mute all sounds from Dragon Chat")}
           </div>
           <div class="setting-control"></div>
         </div>
@@ -562,7 +562,7 @@ export function initGeneralSection({$root}: GeneralSectionProperties): void {
 
   async function factoryResetSettings(): Promise<void> {
     const clearAppDataMessage =
-      "When the application restarts, it will be as if you have just downloaded Zulip app.";
+      "When the application restarts, it will be as if you have just downloaded the Dragon Chat app.";
     const getAppPath = path.join(app.getPath("appData"), app.name);
 
     const {response} = await dialog.showMessageBox({