+++ app/renderer/js/pages/preference/servers-section.ts
@@ -13,7 +13,9 @@ export function initServersSection({$root}: ServersSectionProperties): void {
     <div class="add-server-modal">
       <div class="modal-container">
         <div class="settings-pane" id="server-settings-pane">
-          <div class="page-title">${t.__("Add a Zulip organization")}</div>
+          <div class="page-title">
+            ${t.__("Add a Dragon Chat organization")}
+          </div>
           <div id="new-server-container"></div>
         </div>
       </div>