+++ app/renderer/js/tray.ts
@@ -147,7 +147,7 @@ function sendAction<Channel extends keyof RendererMessage>(
 const createTray = function (): void {
   const contextMenu = Menu.buildFromTemplate([
     {
-      label: "Zulip",
+      label: "Dragon Chat",
       click() {
         ipcRenderer.send("focus-app");
       },