+++ tests/test-add-organization.js
@@ -14,12 +14,9 @@ test("add-organization", async (t) => {
     app.on("window", (win) => put(windows, win));
 
     const mainWindow = await take(windows);
-    t.equal(await mainWindow.title(), "Zulip");
+    t.equal(await mainWindow.title(), "Dragon Chat");
 
-    await mainWindow.fill(
-      ".setting-input-value",
-      "zulip-desktop-test.zulipchat.com",
-    );
+    await mainWindow.fill(".setting-input-value", "chat.dragonplayground.com");
     await mainWindow.click("#connect");
 
     const orgWebview = await take(windows);